Move gripper commands to gripper package
