+++ src/main/java/bhs/devilbotz/commands/gripper/GripperClose.java
@@ -2,28 +2,28 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package bhs.devilbotz.commands;
+package bhs.devilbotz.commands.gripper;
 
 import bhs.devilbotz.subsystems.Gripper;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
-public class GripperIdle extends CommandBase {
-  private Gripper gripper;
+public class GripperClose extends CommandBase {
+  private final Gripper gripper;
 
-  public GripperIdle(Gripper gripper) {
+  public GripperClose(Gripper gripper) {
     this.gripper = gripper;
     addRequirements(gripper);
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {
-    gripper.stop();
-  }
+  public void initialize() {}
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {}
+  public void execute() {
+    gripper.close();
+  }
 
   // Called once the command ends or is interrupted.
   @Override
@@ -32,6 +32,6 @@ public class GripperIdle extends CommandBase {
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return false;
+    return true;
   }
 }
